chore: renames load method
