+++ internal/log/log.go
@@ -1,6 +1,9 @@
 package log
 
-import "log/slog"
+import (
+	"context"
+	"log/slog"
+)
 
 var Logger = slog.Default()
 
@@ -8,3 +11,13 @@ var Logger = slog.Default()
 func SetHandler(handler slog.Handler) {
 	Logger = slog.New(handler)
 }
+
+// Ported from slog.DiscardHandler which makes this available from Go 1.24
+var DiscardHandler slog.Handler = discardHandler{}
+
+type discardHandler struct{}
+
+func (dh discardHandler) Enabled(context.Context, slog.Level) bool  { return false }
+func (dh discardHandler) Handle(context.Context, slog.Record) error { return nil }
+func (dh discardHandler) WithAttrs(attrs []slog.Attr) slog.Handler  { return dh }
+func (dh discardHandler) WithGroup(name string) slog.Handler        { return dh }
