--- log.go
@@ -1,16 +1,0 @@
-package pakay
-
-import (
-	"context"
-	"log/slog"
-)
-
-// Ported from slog.DiscardHandler which makes this available from Go 1.24
-var DiscardHandler slog.Handler = discardHandler{}
-
-type discardHandler struct{}
-
-func (dh discardHandler) Enabled(context.Context, slog.Level) bool  { return false }
-func (dh discardHandler) Handle(context.Context, slog.Record) error { return nil }
-func (dh discardHandler) WithAttrs(attrs []slog.Attr) slog.Handler  { return dh }
-func (dh discardHandler) WithGroup(name string) slog.Handler        { return dh }
