+++ view/secrets_test.go
@@ -35,7 +35,7 @@ func TestListSecrets(t *testing.T) {
       key: TEST_ENV_VAR_3
 `
 
-	err := pakay.LoadSecretsFromBytes([]byte(config))
+	err := pakay.LoadSecretsConfig([]byte(config))
 	require.NoError(t, err)
 
 	ctx := context.Background()
